Introduce CameraDirector interface

+++ src/engine/MouseEventDispatcher.cpp
@@ -10,7 +10,7 @@ namespace {
 
 void mouse_position_callback(GLFWwindow* window, double xpos, double ypos) {
     dbgfln("Mouse position event: xpos=%f, ypos=%f", xpos, ypos);
-    g_mouse_event_dispatcher->dispatch_event({window, xpos, ypos});
+    g_mouse_event_dispatcher->dispatch_event({window, (long)xpos, (long)ypos});
 }
 
 void mouse_button_callback(GLFWwindow* window, int button, int action, int mods) {
